Button + controller/scene addition
Realized it would take maybe two minutes of effort to get the search function to work as it's effectively identical to the last I added.

Added proper fx IDs to the ShowClient screen or whatever I called it.  Yeah, this isn't exactly a big set of changes, but it moves things forwards a little.